+++ v3/errors.go
@@ -41,15 +41,25 @@ package pipe
 
 import "fmt"
 
-// ErrPipelineNonZeroStatusCode is the error we return when a pipeline
-// step returns a non-zero status code AND a nil error
-type ErrPipelineNonZeroStatusCode struct {
-	statusCode int
+// ErrEmptyKey is returned whenever we're given a key that is zero-length
+// or only contains whitespace
+type ErrEmptyKey struct{}
+
+func (e ErrEmptyKey) Error() string {
+	return "zero-length key, or key only contains whitespace"
 }
 
-func (e ErrPipelineNonZeroStatusCode) Error() string {
+// ErrNonZeroStatusCode is the error we return when a sequence has finished
+// with a non-zero status code and no error of its own
+type ErrNonZeroStatusCode struct {
+	sequenceType string
+	statusCode   int
+}
+
+func (e ErrNonZeroStatusCode) Error() string {
 	return fmt.Sprintf(
-		"pipeline exited with non-zero status code %d",
+		"%s exited with non-zero status code %d",
+		e.sequenceType,
 		e.statusCode,
 	)
 }